requirements.txt
discord.py
requests
Pillow
python-dotenv
flask